cambio de tipo de dato en precioUnitario

+++ Application/Dtos/Producto/CrearProductoDto.cs
@@ -6,7 +6,7 @@ namespace Application.Dtos.Producto
     {
 
         required public string Nombre { get; set; }
-        required public double PrecioUnitario { get; set; }
+        required public decimal PrecioUnitario { get; set; }
         required public string Descripcion { get; set; }
 
     }
