+++ Application/Dtos/Producto/RespuestaProductoDto.cs
@@ -6,7 +6,7 @@ namespace Application.Dtos.Producto
     {
 
         public string Nombre { get; set; } = string.Empty;
-        public double PrecioUnitario { get; set; }
+        public decimal PrecioUnitario { get; set; }
         public string Descripcion { get; set; } = string.Empty;
 
     }
